requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
vapi-python==0.1.7
requests==2.31.0
python-dotenv==1.0.0
amadeus==8.1.0

